refs#16
In PARTONS : 
- Update environment configuration script.
- Create new scenario to compute many kinematics with one GPD model.
- Rename SQL file.
- Create SQL file for MySQL database.
- New implementation of ResourceManager (only Scenario supported)
- Remove ResultList<T> class : revert to List<T>.
- Update Computation class : remove association with Scenario and EnvironmentConfiguration classes.
- Update KinematicsUtils : get GPD kinematic from file.
- Remove Computation association from Result class.
- Add copy constructors in many classes.
- In Task class add pointer to Scenario (parent).
- In GPDKinematic : add constructor with string values (value from file).
-  Implement ResultInfo to keep Computation, Scenario and EnvironementConfiguration information for a Result.
- Implement ResultDaoService as a draft to insert batch of results from a temporary CSV dump file to the database with the specific "LOAD DATA INFILE" SQL query.
- Update related DaoService.
- Implement ResultInfoDaoService.
- Modify AutomationService, DefaultXMLParser, XMLParserI to handle the new logical with ResourceManager and Scenario.

+++ include/partons/database/common/dao/ComputationDao.h
@@ -9,7 +9,7 @@
  */
 
 #include <QtSql/qsqlquery.h>
-#include <stddef.h>
+//#include <stddef.h>
 #include <ctime>
 
 #include "../../../BaseObject.h"
@@ -31,8 +31,9 @@ public:
     Computation selectByIndexId(const int indexId) const;
     int getComputationIdByDateTime(const time_t &dateTime) const;
 
-    int insertIntoScenarioComputation(const size_t scenarioTaskIndexNumber,
-            const int scenarioId, const int computationId) const;
+    int insertIntoScenarioComputation(
+            const unsigned int scenarioTaskIndexNumber, const int scenarioId,
+            const int computationId) const;
 
     bool isAvailable(const int computationId) const;
